#68 - feat : [FE] 프론트 플로우에 맞게 수정

+++ frontend/Load_Frontend/lib/views/box_simulation/video_overlay_widget.dart
@@ -34,7 +34,7 @@ class VideoControlsOverlayState extends State<VideoControlsOverlay> {
     setState(() {
       _currentItem = gCurrItem;
       _totalItems = gTotalItems;
-      _currentValue = _currentItem / (_totalItems as double) * 100;
+      _currentValue = _currentItem / _totalItems * 100;
       if (_currentValue > 100)
         _currentValue = 100.0;
       else if (_currentValue < 0) _currentValue = 0.0;
